完成更新词库和优化删除词库逻辑

+++ src/main/java/dictation/word/service/impl/lib/UserLibServiceImpl.java
@@ -6,9 +6,12 @@ import dictation.word.dao.lib.UserLibMapper;
 import dictation.word.entity.lib.tables.Lib;
 import dictation.word.entity.lib.tables.UserLib;
 import dictation.word.exception.DelException;
+import dictation.word.exception.IllegalDataException;
 import dictation.word.exception.NoPermissionException;
 import dictation.word.service.i.lib.LibService;
+import dictation.word.service.i.lib.LibWordService;
 import dictation.word.service.i.lib.UserLibService;
+import dictation.word.service.i.word.WordService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -18,17 +21,33 @@ import javax.annotation.Resource;
 public class UserLibServiceImpl extends ServiceImpl<UserLibMapper, UserLib> implements UserLibService {
     @Resource
     LibService libService;
+    @Resource
+    LibWordService libWordService;
+    @Resource
+    WordService wordService;
 
     @Override
     @Transactional(rollbackFor = Exception.class)
     public boolean removeLib(int libId, int userId) {
+        if (libId == 1) {
+            throw new IllegalDataException("不能删除根词库");
+        }
         remove(Wrappers.<UserLib>lambdaQuery()
                 .eq(UserLib::getUserId, userId)
                 .eq(UserLib::getLibId, libId));
         int uses = count(Wrappers.<UserLib>lambdaQuery()
                 .eq(UserLib::getLibId, libId));
         if (uses == 0) {
+            //删除库
             if (!libService.removeById(libId)) {
+                throw new DelException("删除失败");
+            }
+            //删除库的单词
+            if (!libWordService.removeByLibId(libId)) {
+                throw new DelException("删除失败");
+            }
+            //删除自定义释义
+            if (!wordService.removeByLibId(libId)) {
                 throw new DelException("删除失败");
             }
         }
